# src/tickethub/routers/tickets.py
from typing import Annotated

from fastapi import APIRouter, Depends, HTTPException, Query, Request, status
from sqlalchemy.ext.asyncio import AsyncSession

from tickethub.cache import delete_cache_pattern, get_cache, set_cache
from tickethub.database import get_db_session
from tickethub.models import Ticket
from tickethub.rate_limit import limiter
from tickethub.repositories.tickets import (
    create_ticket,
    get_ticket_by_id,
    list_tickets,
    search_tickets,
    update_ticket,
)
from tickethub.routers.auth import get_current_user
from tickethub.schemas import (
    AuthUserResponse,
    TicketCreate,
    TicketDetail,
    TicketListItem,
    TicketListResponse,
    TicketPriority,
    TicketStatus,
    TicketUpdate,
)

router = APIRouter(prefix="/tickets", tags=["tickets"])


def to_ticket_list_item(ticket: Ticket) -> TicketListItem:
    # Skraćuje opis za prikaz u listi
    description = ticket.description

    if description is not None and len(description) > 100:
        description = description[:100]

    return TicketListItem(
        id=ticket.id,
        title=ticket.title,
        status=ticket.status,
        priority=ticket.priority,
        description=description,
    )


@router.get("", response_model=TicketListResponse)
@limiter.limit("60/minute")
async def get_tickets(
    request: Request,
    status_filter: TicketStatus | None = Query(default=None, alias="status"),
    priority_filter: TicketPriority | None = Query(default=None, alias="priority"),
    limit: int = Query(default=20, ge=1, le=100),
    offset: int = Query(default=0, ge=0),
    session: AsyncSession = Depends(get_db_session),
) -> TicketListResponse:
    # Vraća paginiranu listu ticketa iz lokalne baze uz cache
    cache_key = (
        "tickets:list:"
        f"status={status_filter.value if status_filter else ''}:"
        f"priority={priority_filter.value if priority_filter else ''}:"
        f"limit={limit}:"
        f"offset={offset}"
    )

    cached_data = await get_cache(cache_key)

    if cached_data is not None:
        return TicketListResponse.model_validate(cached_data)

    tickets, total = await list_tickets(
        session=session,
        status=status_filter.value if status_filter else None,
        priority=priority_filter.value if priority_filter else None,
        limit=limit,
        offset=offset,
    )

    response = TicketListResponse(
        total=total,
        limit=limit,
        offset=offset,
        items=[to_ticket_list_item(ticket) for ticket in tickets],
    )

    await set_cache(cache_key, response.model_dump(mode="json"))

    return response


@router.get("/search", response_model=TicketListResponse)
@limiter.limit("60/minute")
async def search_tickets_endpoint(
    request: Request,
    q: str = Query(min_length=1, max_length=100, pattern=r".*\S.*"),
    limit: int = Query(default=20, ge=1, le=100),
    offset: int = Query(default=0, ge=0),
    session: AsyncSession = Depends(get_db_session),
) -> TicketListResponse:
    # Pretražuje tickete po nazivu uz cache
    cache_key = f"tickets:search:q={q}:limit={limit}:offset={offset}"

    cached_data = await get_cache(cache_key)

    if cached_data is not None:
        return TicketListResponse.model_validate(cached_data)

    tickets, total = await search_tickets(
        session=session,
        q=q,
        limit=limit,
        offset=offset,
    )

    response = TicketListResponse(
        total=total,
        limit=limit,
        offset=offset,
        items=[to_ticket_list_item(ticket) for ticket in tickets],
    )

    await set_cache(cache_key, response.model_dump(mode="json"))

    return response


@router.get("/{ticket_id}", response_model=TicketDetail)
@limiter.limit("60/minute")
async def get_ticket_detail(
    request: Request,
    ticket_id: int,
    session: AsyncSession = Depends(get_db_session),
) -> TicketDetail:
    # Vraća detalje jednog ticketa
    ticket = await get_ticket_by_id(session=session, ticket_id=ticket_id)

    if ticket is None:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Ticket nije pronađen.",
        )

    return TicketDetail.model_validate(ticket)


@router.post("", response_model=TicketDetail, status_code=status.HTTP_201_CREATED)
@limiter.limit("60/minute")
async def create_ticket_endpoint(
    request: Request,
    current_user: Annotated[AuthUserResponse, Depends(get_current_user)],
    ticket_data: TicketCreate,
    session: AsyncSession = Depends(get_db_session),
) -> TicketDetail:
    # Kreira novi ticket i briše zastarjeli cache
    ticket = await create_ticket(
        session=session,
        ticket_data=ticket_data,
    )

    await session.commit()
    await session.refresh(ticket)

    await delete_cache_pattern("tickets:*")
    await delete_cache_pattern("stats:*")

    return TicketDetail.model_validate(ticket)


@router.patch("/{ticket_id}", response_model=TicketDetail)
@limiter.limit("60/minute")
async def update_ticket_endpoint(
    request: Request,
    ticket_id: int,
    ticket_data: TicketUpdate,
    current_user: Annotated[AuthUserResponse, Depends(get_current_user)],
    session: AsyncSession = Depends(get_db_session),
) -> TicketDetail:
    # Mijenja postojeći ticket i briše zastarjeli cache
    ticket = await get_ticket_by_id(session=session, ticket_id=ticket_id)

    if ticket is None:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Ticket nije pronađen.",
        )

    updated_ticket = await update_ticket(
        session=session,
        ticket=ticket,
        ticket_data=ticket_data,
    )

    await session.commit()
    await session.refresh(ticket)

    await delete_cache_pattern("tickets:*")
    await delete_cache_pattern("stats:*")

    return TicketDetail.model_validate(updated_ticket)